# Update subset to handle large number of selectors better