@nalimilan - could you please have a look. When we finalize the description I will add the same for `subset!`.

The reason for this note is that we are using recurrence to calculate row condition (so when there are many columns we get stack overflow due too deep recurrence and in general very slow compilation)