@bkamins 2022-01-13
Alternatively I can change the implementation to use other method of computation of the condition when very many conditions are passed - what do you think would be better?