@bkamins 2022-01-18
@nalimilan - do you have an opinion which approach (fixing docstring or changing the implementation) to take? (after thinking I start to lean towards faster implementation, but it would complicate code - as usual)