@nalimilan 2022-01-19
As you prefer. If that's not too hard to handle, switching automatically to the fastest implementation would make sense.