@bkamins 2022-01-19
I have added a new code path for more than 16 conditions.

After adding a new code path the timings are as follows.

### Current release

```
julia> @testset "wide" begin
           df = DataFrame(a=1:10);
               for i in 1:20
               @time @test subset(df, [:a => ByRow(x -> true) for _ in 1:i]) == df
           end
       end
  1.379549 seconds (2.91 M allocations: 165.628 MiB, 9.93% gc time, 99.86% compilation time)
  0.165773 seconds (411.50 k allocations: 20.766 MiB, 4.46% gc time, 99.65% compilation time)
  0.156739 seconds (428.29 k allocations: 21.115 MiB, 99.54% compilation time)
  0.138686 seconds (438.45 k allocations: 21.079 MiB, 5.63% gc time, 99.57% compilation time)
  0.158869 seconds (523.83 k allocations: 25.285 MiB, 99.55% compilation time)
  0.152105 seconds (527.79 k allocations: 24.723 MiB, 4.64% gc time, 99.54% compilation time)
  0.169468 seconds (580.98 k allocations: 26.839 MiB, 4.11% gc time, 99.63% compilation time)
  0.169140 seconds (638.20 k allocations: 29.151 MiB, 99.60% compilation time)
  0.201163 seconds (697.21 k allocations: 31.513 MiB, 3.42% gc time, 99.64% compilation time)
  0.203939 seconds (758.28 k allocations: 34.012 MiB, 3.53% gc time, 99.65% compilation time)
  0.218176 seconds (820.98 k allocations: 36.483 MiB, 3.57% gc time, 99.67% compilation time)
  0.243496 seconds (885.31 k allocations: 39.140 MiB, 3.36% gc time, 99.72% compilation time)
  0.251012 seconds (953.46 k allocations: 41.728 MiB, 99.71% compilation time)
  0.292917 seconds (1.02 M allocations: 44.470 MiB, 4.32% gc time, 97.67% compilation time)
  0.302242 seconds (1.10 M allocations: 47.339 MiB, 2.66% gc time, 99.71% compilation time)
  0.320316 seconds (1.17 M allocations: 50.303 MiB, 2.56% gc time, 99.76% compilation time)
  0.334654 seconds (1.25 M allocations: 53.176 MiB, 2.39% gc time, 99.74% compilation time)
  0.341822 seconds (1.33 M allocations: 56.280 MiB, 2.11% gc time, 99.73% compilation time)
  0.365099 seconds (1.41 M allocations: 59.453 MiB, 4.51% gc time, 99.76% compilation time)
  0.396918 seconds (1.49 M allocations: 62.738 MiB, 2.08% gc time, 99.77% compilation time)

julia> @testset "wide" begin
           df = DataFrame(a=1:10);
               for i in 1:20
               @time @test subset(df, [:a => ByRow(x -> true) for _ in 1:i]) == df
           end
       end
  0.184042 seconds (536.17 k allocations: 29.179 MiB, 99.49% compilation time)
  0.004413 seconds (3.39 k allocations: 206.031 KiB, 94.39% compilation time)
  0.004887 seconds (4.43 k allocations: 265.022 KiB, 94.88% compilation time)
  0.008614 seconds (5.27 k allocations: 302.647 KiB, 96.77% compilation time)
  0.005421 seconds (5.60 k allocations: 317.059 KiB, 94.12% compilation time)
  0.006612 seconds (5.98 k allocations: 334.157 KiB, 94.17% compilation time)
  0.006547 seconds (6.36 k allocations: 356.428 KiB, 94.30% compilation time)
  0.006940 seconds (6.75 k allocations: 374.948 KiB, 94.48% compilation time)
  0.007043 seconds (7.15 k allocations: 397.906 KiB, 94.62% compilation time)
  0.007847 seconds (7.53 k allocations: 416.735 KiB, 93.52% compilation time)
  0.007711 seconds (7.92 k allocations: 435.554 KiB, 94.70% compilation time)
  0.009285 seconds (8.30 k allocations: 452.763 KiB, 95.30% compilation time)
  0.008392 seconds (8.69 k allocations: 469.269 KiB, 94.86% compilation time)
  0.009730 seconds (9.07 k allocations: 488.243 KiB, 95.04% compilation time)
  0.009656 seconds (9.45 k allocations: 504.687 KiB, 94.87% compilation time)
  0.011013 seconds (9.84 k allocations: 522.536 KiB, 95.38% compilation time)
  0.010529 seconds (10.22 k allocations: 538.917 KiB, 95.21% compilation time)
  0.010824 seconds (10.60 k allocations: 556.923 KiB, 94.74% compilation time)
  0.012236 seconds (10.99 k allocations: 573.388 KiB, 95.31% compilation time)
  0.022727 seconds (11.42 k allocations: 591.603 KiB, 47.75% gc time, 97.10% compilation time)
```

### After this PR

```
julia> @testset "wide" begin
           df = DataFrame(a=1:10);
               for i in 1:20
               @time @test subset(df, [:a => ByRow(x -> true) for _ in 1:i]) == df
           end
       end
  1.357905 seconds (3.02 M allocations: 171.580 MiB, 5.09% gc time, 99.88% compilation time)
  0.246806 seconds (411.52 k allocations: 20.770 MiB, 33.45% gc time, 99.76% compilation time)
  0.158856 seconds (428.30 k allocations: 21.116 MiB, 99.57% compilation time)
  0.209060 seconds (438.45 k allocations: 21.082 MiB, 35.55% gc time, 99.57% compilation time)
  0.160563 seconds (523.83 k allocations: 25.290 MiB, 99.56% compilation time)
  0.168839 seconds (527.79 k allocations: 24.717 MiB, 5.05% gc time, 99.25% compilation time)
  0.181965 seconds (580.98 k allocations: 26.841 MiB, 4.27% gc time, 99.59% compilation time)
  0.182608 seconds (638.20 k allocations: 29.148 MiB, 99.56% compilation time)
  0.203064 seconds (697.21 k allocations: 31.512 MiB, 3.37% gc time, 99.65% compilation time)
  0.222555 seconds (758.28 k allocations: 34.014 MiB, 99.65% compilation time)
  0.226757 seconds (820.97 k allocations: 36.478 MiB, 4.66% gc time, 99.65% compilation time)
  0.247002 seconds (885.30 k allocations: 39.127 MiB, 4.11% gc time, 99.69% compilation time)
  0.263710 seconds (953.46 k allocations: 41.746 MiB, 3.24% gc time, 99.65% compilation time)
  0.291812 seconds (1.02 M allocations: 44.465 MiB, 3.36% gc time, 99.71% compilation time)
  0.316660 seconds (1.10 M allocations: 47.336 MiB, 3.01% gc time, 99.73% compilation time)
  0.334900 seconds (1.17 M allocations: 50.294 MiB, 2.60% gc time, 99.73% compilation time)
  0.234488 seconds (681.70 k allocations: 33.238 MiB, 3.34% gc time, 99.48% compilation time)
  0.010446 seconds (10.58 k allocations: 555.281 KiB, 95.10% compilation time)
  0.011420 seconds (10.97 k allocations: 571.620 KiB, 95.64% compilation time)
  0.011522 seconds (11.40 k allocations: 589.631 KiB, 95.25% compilation time)

julia> @testset "wide" begin
           df = DataFrame(a=1:10);
               for i in 1:20
               @time @test subset(df, [:a => ByRow(x -> true) for _ in 1:i]) == df
           end
       end
  0.193705 seconds (536.18 k allocations: 29.236 MiB, 7.70% gc time, 99.57% compilation time)
  0.004700 seconds (3.39 k allocations: 206.031 KiB, 94.45% compilation time)
  0.005274 seconds (4.43 k allocations: 265.022 KiB, 92.49% compilation time)
  0.008448 seconds (5.27 k allocations: 302.835 KiB, 96.56% compilation time)
  0.005928 seconds (5.60 k allocations: 317.059 KiB, 94.29% compilation time)
  0.006156 seconds (5.98 k allocations: 334.157 KiB, 94.65% compilation time)
  0.007340 seconds (6.36 k allocations: 356.428 KiB, 91.86% compilation time)
  0.006614 seconds (6.75 k allocations: 374.948 KiB, 94.35% compilation time)
  0.006861 seconds (7.15 k allocations: 397.906 KiB, 94.63% compilation time)
  0.007529 seconds (7.53 k allocations: 416.735 KiB, 94.53% compilation time)
  0.007730 seconds (7.92 k allocations: 435.554 KiB, 93.93% compilation time)
  0.008697 seconds (8.30 k allocations: 452.763 KiB, 95.25% compilation time)
  0.008500 seconds (8.69 k allocations: 469.269 KiB, 94.95% compilation time)
  0.009921 seconds (9.07 k allocations: 488.243 KiB, 94.84% compilation time)
  0.010125 seconds (9.45 k allocations: 504.687 KiB, 95.02% compilation time)
  0.010552 seconds (9.84 k allocations: 522.536 KiB, 95.36% compilation time)
  0.010386 seconds (10.20 k allocations: 537.448 KiB, 95.38% compilation time)
  0.010849 seconds (10.58 k allocations: 555.267 KiB, 95.43% compilation time)
  0.011165 seconds (10.96 k allocations: 571.606 KiB, 94.90% compilation time)
  0.011723 seconds (11.39 k allocations: 589.618 KiB, 95.46% compilation time)
```
So 16 seems to be reasonable (after compilation)